move freetype dependency to draw2dimg
